pyproject.toml
[project]
name = "feishubookkeeping"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "lark-oapi>=1.5.3",
    "pyyaml>=6.0.3",
]

[[tool.uv.index]]
url = "https://mirrors.aliyun.com/pypi/simple"
default = true
